leveraging a skeleton
functional components separate for easier development
