+++ logs/README.md
@@ -1,0 +1,1 @@
+Your Slim Framework application's log files will be written to this directory.
