+++ src/dependencies.php
@@ -1,0 +1,23 @@
+<?php
+// DIC configuration
+
+// Get container
+$container = $app->getContainer();
+
+// view renderer
+$container['renderer'] = function ($c) {
+    $settings = $c->get('settings')['renderer'];
+
+    return new Slim\Views\PhpRenderer($settings['template_path']);
+};
+
+// monolog
+$container['logger'] = function ($c) {
+    $settings = $c->get('settings')['logger'];
+    $logger = new Monolog\Logger($settings['name']);
+    $logger->pushProcessor(new Monolog\Processor\UidProcessor());
+    $logger->pushHandler(new Monolog\Handler\StreamHandler('./logs/'.$settings['name'].'.log', Monolog\Logger::DEBUG));
+    $logger->pushHandler(new Monolog\Handler\StreamHandler('php://stdout', Monolog\Logger::DEBUG)); // <<< uses a stream
+
+    return $logger;
+};
