+++ src/middleware.php
@@ -1,0 +1,4 @@
+<?php
+// Application middleware
+
+// e.g: $app->add(new \Slim\Csrf\Guard);
