--- tests/Foo.php
@@ -1,7 +1,0 @@
-<?php
-class Foo
-{
-    public function __construct()
-    {
-    }
-}
