--- tests/README
@@ -1,18 +1,0 @@
-Slim Framework Unit Tests
-
-Follow the directions below to run the Slim Framework unit tests. You'll need the latest version of PHPUnit. To save development time, these unit tests require PHP >= 5.3. However, the Slim Framework itself requires only PHP >= 5.2.
-
-1. Install the latest version of PHPUnit
-Visit http://www.phpunit.de/ for installation instructions.
-
-2. Run PHPUnit
-From the filesystem directory that contains the `tests` directory, you may run all unit tests or specific unit tests. Here are several examples. The '$>' in the examples below is your command prompt.
-
-To run all tests:
-$> phpunit tests
-
-To run all HTTP-related tests:
-$> phpunit tests/Http
-
-To run only the HTTP Request tests:
-$> phpunit tests/Http/RequestTest
