--- tests/templates/test.php
@@ -1,1 +1,0 @@
-test output <?php echo $foo; ?>
